hospital_A/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
psycopg2-binary   # postgres
dotenv
